pyproject.toml
[tool.poetry]
name = "acq-btf-with-aries"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aries-python = {git = "https://github.com/2-propanol/aries_python"}
easypyspin = {git = "https://github.com/elerac/EasyPySpin"}
opencv-python = "^4.4.0"
numpy = "^1.19.2"
spinnaker-python = {path = "spinnaker_python-2.0.0.147-cp38-cp38-win_amd64.whl"}
scipy = "^1.5.2"
tqdm = "^4.50.0"
pandas = "^1.1.2"
colour-demosaicing = "^0.1.5"
nptyping = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ptpython = "^3.0.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
